+++ packages/ui/src/components/index.ts
@@ -1,2 +1,4 @@
 export * from "./Button";
 export * from "./Callout";
+export * from "./Chip";
+export * from "./Tag";