+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sopt-mds/ui
 
+## 1.0.3
+
+### Patch Changes
+
+- c1aca84: register chip component
+
+## 1.0.2
+
+### Patch Changes
+
+- 99b1018: register tag component
+
 ## 1.0.1
 
 ### Patch Changes